separating types into modules
