+++ crates/scouter_settings/Cargo.toml
@@ -13,7 +13,9 @@ description = "Client and server contract for scouter"
 
 [dependencies]
 scouter-error = { workspace = true }
+scouter-types = { workspace = true }
 base64 = { workspace = true }
+pyo3 = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 tracing = { workspace = true }
