+++ crates/scouter_settings/src/auth.rs
@@ -1,0 +1,7 @@
+use serde::Serialize;
+
+#[derive(Debug, Clone, Default, Serialize)]
+pub struct AuthSettings {
+    pub jwt_secret: String,
+    pub refresh_secret: String,
+}
