+++ crates/scouter_types/src/lib.rs
@@ -4,14 +4,18 @@ pub mod custom;
 pub mod dispatch;
 pub mod drift;
 pub mod feature;
+pub mod http;
 pub mod psi;
 pub mod records;
 pub mod spc;
+pub mod storage;
 pub mod util;
 
 pub use cron::*;
 pub use dispatch::*;
 pub use drift::*;
 pub use feature::*;
+pub use http::*;
 pub use records::*;
+pub use storage::*;
 pub use util::*;
