+++ crates/scouter_types/src/storage.rs
@@ -1,0 +1,9 @@
+use serde::Serialize;
+
+#[derive(Debug, PartialEq, Clone, Serialize)]
+pub enum StorageType {
+    Google,
+    Aws,
+    Local,
+    Azure,
+}
